backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
einops==0.7.0
fast-pytorch-kmeans==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
tqdm==4.66.1
python-json-logger==2.0.7
